Move top panel into separate struct
Moved the top panel into a separate struct and file to further separate
functionality from presentation.

Notice that the app still needs to set values in the panel;
Notice that it currently is not possible to move the central panel into a
separate struct because the panel needs to access the views in the app struct.