Move RFD "open file" dialog into separate module (#9)
With this change we can move modal logic from the `crate::app` module into a
dedicated module. `crate::app` should be the glue that holds our application
components together, it should not contain such specific implementation details.

Also injected `image_service` dependency into remaining viewmodels so that
these can trigger image loading without going through `crate::app`.

Plan for the future: if we turn the VMs into actor models we can simplify the
modals in a way that the view does not need to poll the promise anymore.
Once the VM is an isolated actor and listens on a channel, we can instruct the
modal to send a message to the VM once the dialog is completed.